requirements.txt
boto3
pandas
slack_sdk
python-dotenv
